extract common layout
